use different input method, get rid of the log object

+++ app/actions/account/actions/delete.ts
@@ -2,11 +2,10 @@ import { Database, COLLECTIONS } from "../../../database/database";
 import { EncryptedAccount } from "../../../database/db-types";
 import { readPassword, readConfirm } from "../../../lib/util";
 import { hash } from "../../../lib/crypto";
-import { Log } from "../../../main";
 
 export namespace DeleteAccount {
-	async function tryPasswordOnce(log: Log, email: string, database: Database) {
-		const password = await readPassword(log, 'Please enter the account\'s password');
+	async function tryPasswordOnce(email: string, database: Database) {
+		const password = await readPassword('Please enter the account\'s password');
 
 		const record: Partial<EncryptedAccount> = {
 			email: database.Crypto.dbEncrypt(email),
@@ -20,10 +19,10 @@ export namespace DeleteAccount {
 		return false;
 	}
 
-	async function getPassword(log: Log, email: string, database: Database) {
+	async function getPassword(email: string, database: Database) {
 		for (let i = 0; i < 3; i++) {
-			log.write(`Attempt ${i + 1}/3`);
-			const result = await tryPasswordOnce(log, email, database);
+			console.log(`Attempt ${i + 1}/3`);
+			const result = await tryPasswordOnce(email, database);
 			if (result) {
 				return result;
 			}
@@ -31,38 +30,42 @@ export namespace DeleteAccount {
 		return false;
 	}
 
-	export async function deleteAccount(log: Log, email: string, database: Database) {
-		const record = await getPassword(log, email, database);
+	export async function deleteAccount(email: string, database: Database) {
+		if (!database) {
+			return;
+		}
+
+		const record = await getPassword(email, database);
 		if (record === false) {
-			log.write('Failed 3 times, exiting');
+			console.log('Failed 3 times, exiting');
 			process.exit(1);
 			return;
 		}
 
-		log.write('Deleting user with email', email);
-		await readConfirm(log, 'Are you sure?');
-		await readConfirm(log, 'Are you very very sure?');
+		console.log('Deleting user with email', email);
+		await readConfirm('Are you sure?');
+		await readConfirm('Are you very very sure?');
 
 		const id = record._id;
 
-		log.write('Deleting instances...');
+		console.log('Deleting instances...');
 		//Delete all instances from the instances collection
 		await database.Manipulation.deleteMany(COLLECTIONS.INSTANCES, {
 			user_id: database.Crypto.dbEncrypt(id.toHexString())
 		});
 
-		log.write('Deleting passwords...');
+		console.log('Deleting passwords...');
 		//Delete all passwords from the passwords collection
 		await database.Manipulation.deleteMany(COLLECTIONS.PASSWORDS, {
 			user_id: database.Crypto.dbEncrypt(id.toHexString())
 		});
 
-		log.write('Deleting user record...');
+		console.log('Deleting user record...');
 		//Delete the record from the users collection
 		await database.Manipulation.deleteOne(COLLECTIONS.USERS, {
 			_id: id
 		});
 
-		log.write('Done deleting user with email', email);
+		console.log('Done deleting user with email', email);
 	}
 }
